S3 location in CSV format
This script is written in PySpark and utilizes the AWS Glue service for ETL (Extract, Transform, Load) operations. It appears to be an AWS Glue script that reads data from a raw CSV source in Amazon S3, applies mapping transformations, and writes the transformed data to another S3 location in the Parquet file format.